Change analysis of the body of a predicate (WiP).
The state does hold the current domain of every variable. Before every call
in the body, we check if the call is a valid one with the current state.
Then we calculate the new state after the call and continue to the next one.

The check, if the state is valid, is currently not implemented. Also
we only look at pre specs.